i2c: finished the initial version of the i2c interface.

+++ include/dev/i2c/i2c.h
@@ -80,7 +80,7 @@ struct i2c_message
  */
 struct i2c_client {
 	struct i2c_adapter *adapter; //!< The adapter where it belongs to.
-	uint8_t sla;
+	uint16_t sla;
 	uint32_t freq;
 	
 	/**
